alpha V8.0 almost done!

+++ app/src/main/java/com/example/miller/parkingkoriv4/PrintUtil/PrintUtil.java
@@ -85,6 +85,6 @@ public class PrintUtil {
                 res = (byte) 0xff;
                 break;
         }
-        return (byte) res;
+        return res;
     }
 }
